before applying several refactorings
