+++ README.md
@@ -33,3 +33,4 @@ for very beginnig you can try `-h`.
 ### Requirements
 
 - gitpython should be installed
+
