+++ code_observer.py
@@ -2,19 +2,18 @@ import os
 import ast
 import logging
 from constants import FILES_AMOUNT
-# from itertools import izip
 import argument_parser
 from support_methods import *
 
 logging.basicConfig(level=logging.INFO)
 
 
-def find_files_by_extention(from_path, extention):
+def find_files_by_extention(from_path, extention, amount):
     files_list = []
     for whole_path, dirs, files in os.walk(from_path, topdown=True):
         for file in files:
             files_list.append(extention_only(file, whole_path, extention))
-            if len(files_list) >= FILES_AMOUNT:
+            if len(files_list) >= amount:
                 break
     files_list = filter(None, files_list)
     logging.info('Total found *.%s files amount is: %s' % (extention, len(files_list)))
@@ -39,7 +38,7 @@ def get_trees(files):
 
 def run(args=argument_parser.args):
     location = location_determining(args.source, args.path)
-    files = find_files_by_extention(location, args.extention)
+    files = find_files_by_extention(location, args.extention, args.amount)
     if args.entities == 'functions':
         entity = node_names(get_trees(files))
     else:
